requirements.txt
# Main
Django
djangorestframework
psycopg2-binary
gunicorn
celery[redis]
django-environ
django_redis

# Dev
pylint
flake8
ruff
mypy
hypothesis
django-debug-toolbar
redis
drf-yasg
isort
black